+++ src/atcoder/abc2xx/abc208/b_factorial_yen_coin.rs
@@ -1,4 +1,6 @@
-fn fact(num: i32) -> i32 {
+// https://atcoder.jp/contests/abc206/tasks/abc206_b
+
+fn fact(num: usize) -> usize {
     let mut result = 1;
 
     for i in 1..=num {
@@ -8,8 +10,7 @@ fn fact(num: i32) -> i32 {
     result
 }
 
-#[allow(dead_code)]
-pub fn run(input: i32) -> i32 {
+pub fn run(input: usize) -> usize {
     let mut total = input;
     let mut result = 0;
