+++ src/atcoder/abc2xx/abc231/a_water_pressure.rs
@@ -1,6 +1,6 @@
 // https://atcoder.jp/contests/abc231/tasks/abc231_a
 
-pub fn run(d: i32) -> f64 {
+pub fn run(d: usize) -> f64 {
     d as f64 / 100.0
 }
